requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-dotenv==0.17.1
requests==2.25.1
slack-bolt==1.6.1
slack-sdk==3.7.0
SQLAlchemy==1.4.22
urllib3==1.26.5
Werkzeug==2.0.1
